feat: ♻ refactor everything
Create components, events, interface, utils folder and change the import statement
