+++ events/event_listener.py
@@ -1,0 +1,10 @@
+class EventListener():
+    def __init__(self) -> None:
+        pass
+    def run(self, event, game):
+        """
+        Oriented Object Observer
+        :param Event event: the triggered event.
+        :param Game game: the game instance.
+        """
+        pass
